requirements.txt
discord.py==2.3.2
requests==2.32.5
flask==3.0.0
supabase>=2.3.0
python-dotenv==1.0.0